requirements.txt
deepseekr==0.1.0
pywebview==4.3.3